Src/requirements.txt
amulet-core==1.7.5
amulet-nbt==1.2.0
specklepy==2.6.7
